feat: added top songs for artist + refactor artist detail to use view model
