+++ flo/Resources/Localizable.xcstrings
@@ -1003,6 +1003,9 @@
         }
       }
     },
+    "Play Top Songs" : {
+
+    },
     "Player color" : {
       "localizations" : {
         "en" : {
